+++ src/main/java/insane96mcp/iguanatweaksexpanded/module/experience/enchanting/EnchantingFeature.java
@@ -176,7 +176,7 @@ public class EnchantingFeature extends JsonFeature {
 
         if (event.getItemStack().is(Items.ENCHANTED_BOOK)) {
             for (Map.Entry<Enchantment, Integer> enchantment : EnchantmentHelper.getEnchantments(event.getItemStack()).entrySet()) {
-                if (enchantment.getKey().isTreasureOnly()) {
+                if (enchantment.getKey().isTreasureOnly() && !enchantment.getKey().isCurse()) {
                     event.getToolTip().add(Component.empty());
                     event.getToolTip().add(Component.translatable("iguanatweaksexpanded.apply_to_enchanting_table").withStyle(ChatFormatting.GREEN));
                     break;