+++ changelog.md
@@ -4,6 +4,7 @@
 * Fixed Solium moss not dropping with durium shears
 * Fixed missing durium shears recipe
 * Fixed Keego items being repairable with Gold
+* Fixed souls steel and durium items amount wrong
 
 ## 1.3.1
 * Added Quaron Reinforced Fishing rod