+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Reduced max hostile mobs in Winter
 * Fixed Solium moss not dropping with durium shears
 * Fixed missing durium shears recipe
 * Fixed Keego items being repairable with Gold