+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## Upcoming
+* Reduced chance for fisher zombies
 * Reduced max hostile mobs in Winter and slightly increased in summer
 * Fixed Solium moss not dropping with durium shears
 * Fixed missing durium shears recipe